+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 - Fixed `form`/`mold` for `binary!` values with an offset.
 - Added `clear` action.
 - Added `remove` action.
-- Added `insert` action for `any-list!`, `any-path!`, `any-string!`.
+- Added `insert` action for `any-list!`, `any-path!`, and `any-string!`.
 - Added `pad` and `repend` helper functions.
 - Redid `email!`'s implementation.
+- Finished implementing `append` for `any-list!`, `any-path!`, and `any-string!`.
 
 
 # Version 0.1.1